repaired bug for_all_elements_in_matrix3d
git-svn-id: file:///home/roberto/Git/svn@4320 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: b4986a0150bb6518b6004bb45bfa154b9de0048b

+++ trunk/xmipp/libraries/data/matrix3d.h
@@ -89,9 +89,9 @@ void applyGeometryBSpline(Matrix3D<T>& V2, const Matrix2D< double > &A,
  * @endcode
  */
 #define FOR_ALL_ELEMENTS_IN_MATRIX3D(V) \
-    for (size_t k=STARTINGZ(V); k<=FINISHINGZ(V); k++) \
-        for (size_t i=STARTINGY(V); i<=FINISHINGY(V); i++) \
-            for (size_t j=STARTINGX(V); j<=FINISHINGX(V); j++)
+    for (int k=STARTINGZ(V); k<=FINISHINGZ(V); k++) \
+        for (int i=STARTINGY(V); i<=FINISHINGY(V); i++) \
+            for (int j=STARTINGX(V); j<=FINISHINGX(V); j++)
 
 /** For all elements in the array between corners.
  * @ingroup VolumesSizeShape
